control-label class for Vertical forms
For Vertical forms, Labels are not colored with green (success) or red (error) css styles since <label> tags do not have control-label class when using labelOptions in field() function. Without 'labelOptions' everything is working. 
Example:
$form->field(
            $model, 
            "attribute", 
            [
                    'labelOptions' => [
                        'label' => $productPercentage->product->product_name
                    ]
                ]
            )->textInput() 

I'm not sure if this is the correct place for fix. But it resolves my problem.